pyproject.toml
[project]
name = "developer-portfolio-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gcloud>=0.18.3",
    "google-adk>=1.18.0",
    "google-generativeai>=0.8.5",
    "requests>=2.32.5",
]
